Backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.20
databases
python-jose==3.3.0
passlib==1.7.4
PyJWT
python-decouple
python-multipart==0.0.6
requests
bcrypt==4.0.1
python-dotenv==1.0.0